pyproject.toml
[project]
name = "mcp-gsc-server"
version = "0.1.0"
description = "Creación de un script para conectarse a través de un MCP a Google Search Console a a nivel de servidor"
readme = "README.md"
author = "Javi Lázaro"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.64.0",
    "google-api-python-client>=2.179.0",
    "google-oauth2-tool>=0.0.3",
    "mcp>=1.13.0",
    "python-dotenv>=1.1.1",
    "typer>=0.16.1",
]
